rename pytest directories and move test util to correct dir
