+++ tests_pytest/common_tests/core/common/collectors/test_weighted_histogram_collector.py
