+++ tests_pytest/fw_tests_common_base/base_tpc_attach2fw_test.py
