+++ tests_pytest/keras_tests/core/test_data_util.py
