+++ tests_pytest/pytorch_tests/target_platform_capabilities/__init__.py
